requirements.txt
Django
pep8
python-dotenv
pillow
django-debug-toolbar
openai